Switched from double to float as the main data-type; added some basic vectorization to common matrix math

+++ DotML.Benchmark/benchmarks/Layers/Convolutional.cs
@@ -15,18 +15,18 @@ public class BenchmarkConvolutionalLayer {
     [Params(32, 64, 128, 256, 512, 1024)]
     public int DIM_LENGTH {get; set;}
 
-    private BatchedFeatureSet<double> input;
-    private BatchedFeatureSet<double> output;
+    private BatchedFeatureSet<float> input;
+    private BatchedFeatureSet<float> output;
     private ConvolutionLayer layer;
 
     [GlobalSetup]
     public void Setup() {
         var layer = new ConvolutionLayer(input_size: new Shape3D(3, DIM_LENGTH, DIM_LENGTH), padding: Padding.Same, stride: 1, filters: ConvolutionFilter.Make(1, OUT_CLASSES, KERNEL));
-        var input = new FeatureSet<double>(layer.InputShape);
-        var output = new FeatureSet<double>(layer.OutputShape);
+        var input = new FeatureSet<float>(layer.InputShape);
+        var output = new FeatureSet<float>(layer.OutputShape);
 
-        this.input = new BatchedFeatureSet<double>(input);
-        this.output = new BatchedFeatureSet<double>(output);
+        this.input = new BatchedFeatureSet<float>(input);
+        this.output = new BatchedFeatureSet<float>(output);
         this.layer = layer;
     }
