+++ DotML.Benchmark/benchmarks/Layers/LayerNorm.cs
@@ -13,18 +13,18 @@ public class BenchmarkLayerNorm {
     [Params(32, 64, 128, 256, 512, 1024, 2048)]
     public int DIM_LENGTH {get; set;}
 
-    private BatchedFeatureSet<double> input;
+    private BatchedFeatureSet<float> input;
     private LayerNorm layer;
 
     [GlobalSetup]
     public void Setup() {
         // Create random dataset
-        input = new BatchedFeatureSet<double>(new Shape4D(1, IMG_CHANNELS, DIM_LENGTH, DIM_LENGTH));
+        input = new BatchedFeatureSet<float>(new Shape4D(1, IMG_CHANNELS, DIM_LENGTH, DIM_LENGTH));
         Random rng = new Random();
         for (var b = 0; b < input.Batches; b++) { 
             for (var c = 0; c < input.Channels; c++) {
                 var mtx = input[b, c];
-                mtx.Apply((el) => rng.NextDouble());
+                mtx.Apply((el) => (float)rng.NextDouble());
             }
         }
         // Init layer
