+++ DotML.Cli/src/IDecoder.cs
@@ -1,7 +1,7 @@
 namespace DotML.Cli;
 
 public interface IDecoder {
-    public IDecodedResult Decode(BatchedFeatureSet<double> output_values);
+    public IDecodedResult Decode(BatchedFeatureSet<float> output_values);
 }
 
 public interface IFileOnlyDecoder : IDecoder {
