+++ DotML.Cli/src/IEmbedder.cs
@@ -3,6 +3,6 @@ using DotML.Network;
 namespace DotML.Cli;
 
 public interface IEmbedder {
-    public BatchedFeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, IEnumerable<FileInfo> files);
-    public BatchedFeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, string raw);
+    public BatchedFeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, IEnumerable<FileInfo> files);
+    public BatchedFeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, string raw);
 }
